TurrisLite: Add support for SERDES switching
Turris Lite has both PEX and mSATA interface connected to one SERDES
on one PCIe slot (no. 0). The pinout is being switched by HW but the
initialization has to be done in U-Boot according to a signal from
the switch that is being sensed by MCU and the bit is available over
I2C.

Ask MCU over I2C during the start for the SATA bit and init SERDES 0
accordingly. Therefore reset is needed when SATA or PCIe card has been
inserted into the slot 0.