BtrFS: Fix path handling problems
Fix strncmp on the lowest boundar in subvolume name: It causes
subvolume with the same name as prefix of another subvolume to be
matched for the full name. I.e. @ matched for @factory.

Fix handling of the slash in the beginning of the path: Leading
slash causes crash when used in btrls or btrload commands with
or without snapshot name.